backend/requirements.txt
asgiref==3.9.2
celery==5.4.0
Django==5.2.6
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.27.2
Pillow==10.4.0
psycopg2-binary==2.9.10
python-slugify==8.0.4
redis==5.2.0
sqlparse==0.5.3
tzdata==2025.2
